move in-memory repos
